subcategory_module

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -86,9 +86,9 @@ class CategoryController extends Controller
     public function destroy(Category $category)
     {
         try {
-            // if($category->subCategories()->exists()) {
-            //     return response()->json(['status' => 'error', 'message' => __('This category has sub categories and can not be deleted')], 402);
-            // }
+            if($category->subCategories()->exists()) {
+                return response()->json(['status' => 'error', 'message' => __('This category has sub categories and can not be deleted')], 402);
+            }
 
             $category->delete();
